Finish classes and tests.
Played a bit with randomizing the diameters to get different averages.

If I were extending this, I would base the number of apples on the
age/height of tree in a more interesting manner. Also, the size of the
apples should be a function of the age of the tree. None of this is
necessary for getting the code to work as requested.